requirements.txt
asyncpg==0.22.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==7.1.2
databases==0.4.1
ecdsa==0.18.0
fastapi==0.63.0
h11==0.14.0
idna==3.4
ormar==0.10.5
psycopg2-binary==2.8.6
pyasn1==0.4.8
pydantic==1.8
python-jose==3.3.0
python-keycloak==2.6.0
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.3.23
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.13
uvicorn==0.13.4
